rls_test: Return if the packet index exceeds the expected list size
The RLS tests will currently fail if the number of received packets exceeds
the expected number of packets. However, the routine that does this check
does not return, which causes an exception to occur when the expected packet
list is indexed into. This patch simply bails out of the routine, saving us
from a spurious exception.

Change-Id: Id8d9c2b6128740a6cb11a22f2c5e0543b96549d9