lib/python/asterisk/pcap: Add a UDP proxy for VoIP packet analysis
This patch adds a UDP proxy for VoIP packet analysis, VOIPProxy. The proxy
will forward received packets to some destination, based on some set of
rules that map a received port to a destination address/port tuple. Default
rules will forward packets from 5060 to 127.0.0.1:5062 and vice versa, with
the proxy listening on 5061.

Other tests may extend the VOIPProxy class or use it to verify received
packets. This patch refactors the existing VOIPListener into two classes:
VOIPListener, which continues to act as an extension of a PcapListener, and
a new base class, VOIPSniffer. The VOIPSniffer class merely provides the
mechanisms for registering callbacks and calling them with parsed packets
(SIPPacket, etc.). Both VOIPProxy and VOIPListener inherit from VOIPSniffer
to use this functionality.

Note that since twisted will return UDP packets to us parsed as a 'string'
object if the packet data is an ASCII string, we can no longer rely on either
obtaining the Layer 2 - Layer 4 headers, nor can we expect to get the UDP data
as binary data. As such, some of the Packet derived classes are now tolerant
to receiving a string, where appropriate. Those classes that are not explicitly
will throw if passed a string because their data should never be of type 'str'.

ASTERISK-25430

Change-Id: I7753a6ab33df88e7634c339adb8532d1b0b35d78