pjsip/sdp_offer_answer/attribute_passthrough: Move files for new formats
The existing attribute_passthrough test attempts to test all attribute format
modules in one shot. While that's nice, new format attribute modules can't be
easily added in that scheme, as the module may not exist for certain versions
of Asterisk. This patch merely reorganizes the tests here by moving the
existing attribute_passthrough test and renaming it as 'speex_h263_h264' -
as those are the attribute modules tested - and changing the
attribute_passthrough folder to be a directory of tests.

Change-Id: I35bb61177d5e0f5be84c32fad9d9fe1e78485707