lib/python/asterisk/pcap: Fix regression introduced by 345bd86237
When 345bd86237 was merged, it assumed that using 'super' to call the base
class initializers would call all instances of that initializer in the
various base classes. Not surprisingly, this isn't how 'super' works at all.
Rather, 'super' guarantees proper method resolution order based on the
appearance of a base class type in the instance's __bases__ attribute. In this
case, because VOIPSniffer implemented __init__, and it appeared prior to
PcapListener in VOIPListener's __bases__, PcapListener was never being
initialized.

Since VOIPListener is a little odd in that it is composing these two types
together, it really does want to explicitly call the __init__ method on
the base types explicitly. As such, this patch changes the usage of 'super'
to use the explicit syntax. With this patch, both base classes have their
initialization methods called, and the various failing tests that still use
a pcap based listener pass again.

Multiple inheritance: proceed with extreme caution.

Change-Id: Ic6b47c505c3f4e73ab120cbcd71d1a8628ca4492