tests/endpoints/statsd: Take into account 'samples' config environment
The build agents run tests using the 'samples' set of configuration files. With
this set of configuration files, an anonymous endpoint will be created, which
will add additional endpoint statistics that the test was not expecting. This
patch accomodates that environment, as it is the expected one for the vast
majority of test execution.

Change-Id: I4e35788c4feeb4354cd42e09857988528808fca5