subscribe-ws: End test when all major events have been received
This test was super unreliable, due to trying to stop the test based on
interleaved events whose ordering is unreliable. Since the generic pluggable
module that manages actions and events:
(a) Has no mechanism to stop a test based on waiting for *all* received
    events, as opposed to *any* received event
(b) Cannot be easily modified to look for all events, due to exposing the
    pluggable module that manages a series of events as opposed to the
    actual events themselves to the actions
This patch opts for using its own pluggable module for the test that looks
for the expected events and, when all are received, stops the test.

Change-Id: If133d8e1c1ea6d686140aca710647c5178da7650