qualify_timeout: Increase qualify timeout.
The qualify_timeout test fails under Jenkins test agents occasionally.
After investigating, this was because the qualify was timing out before
the AMI session was connected. Since the AMI session was not connected,
the required ConnectionStatus event would not get sent, meaning the test
would fail.

By increasing the qualify timeout, we can better ensure that the
qualify_timeout will happen after AMI has been connected, and we will
see the event we expect to see.

Change-Id: If9a011f432e30e406f6dace4c8e2712aedb593b8