lib/python/asterisk/pcap: Fix SDPPacket regression introduced in 43061c6038
A type was introduced in 43061c6038 that only affected SDPPackets generated as
a result of sniffed VoIP traffic. A parameter, ascii_packet, was instead being
referenced as ascii_pack. This patch corrects the spelling of the parameter.

Note that the PJSIP direct media tests were properly able to sniff the traffic
once this was corrected.

Change-Id: I4027736e8cb02b454e0e7e1f11bb876d12d175ac