Added Debug Message for Test Objects Configured as Base Type
When configuring a test module, it is especially easy to
misconfigure the requested test-object as a type of
'test_case.TestCase' instead of 'test_case.TestCaseModule'.
The significant difference between these two types is
'test_case.TestCase' is the base type of the Python test
modules, while 'test_case.TestCaseModule' is a derived type
of 'test_case.TestCase', and serves as the base type for
pluggable modules.

Unfortunately, because 'test_case.TestCaseModule' is a
derived type of 'test_case.TestCase', misconfiguring the
test-object to use the base class, results in the module
successfully loading and appearing to execute succesfully in
a mechanical sense, but fails unexpectedly and without any
meaningful feedback.

This patch attempts to help the developer of test modules
such that if the base type is used, a debug message is piped
to the testsuite log, indicating a possible misconfiguration
of the test-object type in the test-config.yaml.

ASTERISK-25557 Reported By: Ashley Sanders

Change-Id: I8a16d2f43ae5944680ba393f39bdb80f5df79e07