Automatic conversion of tests to use realtime.
This commit adds a pluggable module that can be used to automatically
make a test that currently uses static configuration files get converted
into a realtime test that uses a database.

The conversion works by keeping track of a registry of config files that
it understands how to convert into database data. When one of the known
files is encountered on a test, the configuration is converted to the
database data.

In addition to the conversion, the pluggable module also writes
configuration files necessary to make the test realtime-capable. For
instance, it will write extconfig.conf, res_odbc.conf, and potentially
sorcery.conf files automatically.

This commit only adds support for conversion of pjsip.conf to database,
but the groundwork is laid to be able to write converters that could
convert other types of files as well.

A later commit adds the ability to load pluggable modules globally,
thus allowing for this module to be used widespread, without having
to update individual test-config.yaml files on each realtime-enabled
test.

Change-Id: I3ceaae8b8f4160a6d23f2062ce75017351c1559e