RFC: Add global pluggable modules.
This gives the test runner the ability to load pluggable modules from
the global test-config.yaml in addition to the test-specific
test-config.yaml.

This was done in order to allow test-wide uses of certain pluggable
modules. The module that is added to the example test-config.yaml in
this commit is for the realtime conversion pluggable module. If this
pluggable module is in the test's global configuration, then the module
will be applied to every test that is run.

This change has the "RFC" designation on it mainly because it is based
on another RFC commit. It also has come under scrutiny before and may
still require larger changes than what are presented here.

Change-Id: Ieec1b1cfa48cadab108c4ab65122ce36ab697e4e