RFC: Add runtime option that allows automatic realtime execution.
This adds a command line option to runtests.py that allows for tests to
be run in realtime mode if possible. This makes it so that any tests
that use configuration files that are registered realtime converters
will automatically be run using a realtime database instead.

This change has the "RFC" designation on it because, for one, it's based
on another RFC commit. Second, this feels a bit on the "hackish" side,
so it's up for debate about whether this can be done better.

As it stands, this change has been tested by running tests like:

./runtests.py -r -t tests/channels/pjsip/ami/show_endpoints

and has proven that they get run using realtime rather than the static
pjsip.conf configuration file.

Change-Id: Ieec1b1cfa48cadab108c4ab65122ce36ab697e4e