Add global pluggable modules.
This gives the test runner the ability to load pluggable modules from
the global test-config.yaml in addition to the test-specific
test-config.yaml.

The reason for this is that there are certain pluggable modules that
have the following attributes:
1) They have configuration that is specific to the environment, rather
than to the test itself
2) They cause an overall change in behavior to how a test runs, but the
test itself does not need to be aware of this change.
3) They should be applied to loaded during a large number of tests.

The only pluggable module that fits these criteria at the moment is the
pluggable module that converts a test with static configuration files
into a realtime test. By putting this pluggable module in the global
test-config.yaml file, it allows for tests being run to automatically be
converted into realtime tests without the need for altering any of the
tests themselves.

If a global pluggable module is configured, it will be loaded for all
tests that are run. It is likely, therefore, that test runs that use
global pluggable modules will also specify a certain tag. This allows
for test runs where they can allow tests to opt into the run.

Change-Id: Ieec1b1cfa48cadab108c4ab65122ce36ab697e4e