confbridge_waitmarked_kick: reactor times out too soon
This test failed once and it appears simply to be due to the reactor timing
out. At various points during the AppTest it resets/extends the timeout value.
It is possible on slower systems that the timeout is reached before it can
be extended. This patch increases the test's reactor timeout.

Change-Id: Iab55d34645d3e0039b33945ebe5029111075af74