endpoints/statsd: Fix the expected stats for 'make samples' configuration
When run with a set of configuration files installed by 'make samples', the
endpoints/statsd test will currently fail. This is due to one of the channel
drivers with a 'sample' configuration creating an endpoint in addition to the
endpoints created by the test.

This patch updates the test (again) to track the expected events with the
configuration that the CI infrastructure runs with.

ASTERISK-25137

Change-Id: I704649c4dcf3b72a750fb8db0fa6347700670d4f