cdr/originate-cdr-disposition: Fix ordering of SIP messages in scenario
The SIPp scenario wait-for-call-timeout attempts to send a 487 to a cancelled
INVITE request before responding to the CANCEL request with a 200 OK. As a
result, on faster machines, Asterisk will send the ACK to the 487 prior to
the SIPp scenario sending the 200 OK. This can cause the scenario to fail,
as it will receive the ACK prior to sending the 200.

This patch reorders the 487/200, as Asterisk will not send an ACK until it
receives the final response for the INVITE request. Overall, this makes the
scenario slightly more reliable.

Change-Id: I955b15b488881835b7dd39b6b34cd35c3fb233fd