run-local: Enable reading of setup variables from run-local.conf.
run-local uses environmental variables to control the Asterisk build.
This allows the script to read variables from run-local.conf.  This file
is ignored by git, an example is included as run-local.conf.example.
This change required switching the script from /bin/sh to /bin/bash to
allow variables to be sourced.

Additionally add CONFIGURE_OPTIONS to allow any option to be appended to
the Asterisk ./configure.

Change-Id: I1e2be35889fde93fae3f31189001a7ceba3a6e25