Do not use local Asterisk settings for remote execution.
When attempting to run tests against a remote Asterisk, there can come
issues since the testsuite has some assumptions that the target Asterisk
system is local. Specifically, the Asterisk version, the Asterisk build
options, and the Asterisk modules are all assumed to be gathered from
the system where the testsuite is running.

For the build options and the modules, this change makes it so that we
just always assume the desired build options and modules are present.

What's more interesting is the Asterisk version. runtests.py contains a
-v option that allows for you to manually specify the version. This
change expands that option a bit to make it so that whenever the
AsteriskVersion() is queried, the version passed into runtests.py is
returned. This means that when running against a remote Asterisk, you
MUST specify the -v option in order to tell the testsuite what the
remote Asterisk version is. This will allow for tests to accurately
determine how/if to run.

Change-Id: I1cefd6f3b1ec0a9edb58dec008ddb3461e0949b8