Add option to strip prefix from paths in refleaks-summary.
This adds a '-p' option to specify the prefix that should be stripped
from the paths of all refs.txt files listed in the summary.  This is
used by runtests.py to remove 'logs/', this way the paths listed in
refleaks-summary.txt are relative to that file instead of the testsuite
root.

Change-Id: I6e59325d1e3447a7b3e8fe670f9c262091b62e8e