stasis_bridge: Hangup from dialplan to give time for Stasis messages.
This test terminates based on the hangup of the underlying channels
used during the test. This presents a race condition where the test
will terminate before the Stasis messages have been conveyed over the
ARI WebSocket.

This change makes it so the channels are hung up a period of time after
they are bridged together using Bridge(). This gives ample time for
the Stasis messages to be conveyed.

Change-Id: Icf776149b9d0052b0db5be271608f508d2ee97ff