stasis_bridge: Ensure the test gives time for all events to be received.
This test terminates based on the hangup of the underlying channels
used during the test. This presents a race condition where the test
will terminate before the Stasis messages have been conveyed over the
ARI WebSocket.

This change makes it so an additional channel is originated and waits
for 10 seconds before continuing. This gives ample time for all events
and messages to be conveyed but does not alter the behavior of the
rest of the test.

Change-Id: Icf776149b9d0052b0db5be271608f508d2ee97ff