Fix bouncing sip_bye_also test.
The test was failing for a few reasons

* The SIPp scenarios had no delay between sending an ACK and sending a
  BYE with Also. This meant that sometimes the BYE would be received
  before the SIP channel was bridged, resulting in a failed transfer.
  This was not actually detected by the test as a failure, but I noticed
  it when diagnosing the test.

* The AMI hangup action would sometimes run before all of the calls had
  stabilized. This resulted in Asterisk sending a 503 response to
  incoming INVITEs from SIPp.

Unfortunately, simply increasing the delay before running the AMI action
did not work as expected. The SIPpTestCase would end the test once all
SIPp scenarios had finished. This resulted in a failing test because the
AMI action did not run before the test completed.

This commit introduces two changes to get the test passing.

First, the SIPp scenario that is run now has a two-second delay between
sending the ACK and sending the BYE. This allows time for calls to be
bridged before attempting the transfer.

Second, the SIPpTestCase has been altered to have an option,
"stop-after-scenarios". When set to "False", the test will not stop
after the SIPp scenarios have completed. Stopping the test is left to
another party. To complement this, the SIPpAMIActionTestCase has a new
option, "run-after-scenarios". This makes it so the AMI action is not
attempted until after the SIPp scenarios have all completed.

By using these new options, we can be sure not to run the AMI action to
early, and we can be sure the test won't stop before the AMI action has
a chance to run. The test has been altered to use the HangupMonitor in
order to stop the test.

Change-Id: Iaaa791f24fe52515c42e4c529bb8f0fd6494ea7a