Fix AMI exten_state_list test.
This test would fail from time to time because an out-of-date device
state would be sent when we would request the ExtensionStateList from
AMI.

This was happening because the test would send device state changes and
then immediately request the ExtensionStateList. Unfortunately, the
device state changes had not propagated entirely through the system yet,
and so we were retrieving stale data.

The test has been modified to request the ExtensionStateList after
receiving all ExtensionStatus and PresenceStatus events. This way, we
can be sure the system is completely updated and we are getting timely
data.

Change-Id: I762a614f6bc6bf7cc72ce9710786fadd69f13572