Get sip_bye_also test passing with Asterisk 11.
The test was recently altered to use the HangupMonitor to stop the test
when all channels had hung up. The problem is that in Asterisk 11, blind
transfers are done using channel masquerades. This causes problems
because <ZOMBIE> channels get hung up, and the HangupMonitor has no
record of those channels existing.

The fix is to have the HangupMonitor listen for Rename AMI events. This
way, we can tell when a channel has been renamed to be a zombie variant.
With this change, the test passes.

Change-Id: I1af60d6a0461d974a4269ca3782dd5b85ae619cb