Fix bouncing attended transfer tests.
Within a bridge test, individual runs start when all channels from the
previous run have hung up.

The problem is that sometimes after all channels have hung up and we
decide to move on to the next run, we can get an Unlink event telling us
about the channels we already know to have hung up. The code, seeing
that a Bridge event has occurred, then locks onto the channel names in
that Bridge event. Subsequent events then refer to new channel names, so
the test ignores those events.

The fix here is to only attempt to learn channel names from Link events.
This way, if an Unlink occurs late, the test does not errantly lock onto
the channel names in that Unlink event.

Change-Id: Icf9cc82329ebd3774bf5aa2c4ef76217c430673f