tests/rest_api/channels/originate: Reduce number of channels.
The testsuite can get overwhelmed when 25 channels are originated
at the same time causing its AMI capability to slow down and
disconnect. When this happens it proceeds to originate a second
channel to kick off the whole process again which causes more
problems.

This change merely reduces the number of originated channels
down to 15 so the slower Jenkins nodes can tolerate it.

Change-Id: I51ee075a097dfeef568f0069f8bd22975106c56e