SIP/sip_custom_presence/resubscribe fails due to retransmission(s)
Sporadically, Asterisk takes a little longer than usual to send the 200 OK reply
for the subscribe. When this happens SIPp restransmits the subscribe, however in
this time Asterisk responds to the first subscribe, receives the retransmit and
responds to it as well. The SIPp scenario of course was not expecting another
200 OK response, so fails.

This patch disables retransmits in the test's SIPp scenarios. Asterisk should
always be up and running by the time the scenario starts so should either always
respond at some point or fail altogether.

Change-Id: Ibea55d4cd35d378196e3680e84fc2a537dddf296