pjsip/show_endpoint,pjsip_contact: Add Created as a valid Contact status
Since the sorcery/realtime optimization is accelerating contact qualification,
contact status is reporting Created for these tests instead of Unknown.

For ami/show_endpoint, the Status match regex was updated from
'(Unknown|Reachable)' to '(Unknown|Created|Reachable)'.

For dialplan_functions/pjsip_contact, the dialplan was updated to do a regex
on (Unknown|Created|Reachable) instead of an exact match on Unknown.  rtt was
also updated to a regex on (N/A|[0-9.]+) instead of an exact match on N/A..

Both tests passes with or without the optimization patch.

Change-Id: I0914a2bc8f5289c804e5e11bfd16e7c9e388c012