Fix Failing atxfer_fail_blond test
There were two issues here:

First, the TALK_DETECTED variable was not being set by BackgroundDetect.
This was because the audio being transmitted by Alice was going to Bob
but not being echoed back to Alice. This is fixed by adjusting Bob's
dialplan to use Echo() instead of Wait().

Second, the automatic recall feature for attended transfers was not
working properly for this test. The problem with the recall is that it's
possible to call a specific SIP endpoint, but it's not possible to know
what username to call at that endpoint. So when attempting to recall
Alice, an INVITE gets sent to an Asterisk instance but there is no way
for the Asterisk server to know where in the dialplan to send the call.
This is fixed by simply removing the recall. The test is not intending
to even test the recall, but to ensure that in a failed blond transfer
that all channels are properly hung up.

Change-Id: I8c4175c3abd1cc1d2e71f3a73c3a11ea17436452