pjsip/ami/show_subscriptions: Fill out SIPp scenario more.
In some situations it was possible for the SIPp scenario to end
before the test had really finished, causing Asterisk to shutdown
and the test to fail.

This change fills out the SIPp scenario more to expect another
NOTIFY message and also adds a pause to give the test time to
get inbound subscription details.

Change-Id: I429855159c9f01105c5cc1c111e7c9009fc37d55