tests/apps/mixmonitor: Fix sporadic test failures.
The tests/app/mixmonitor test executes MixMonitor on
some channels and then checks the resulting file size to
ensure they are within tolerance. This fails sometimes.
After looking at the output and the test there is a race
condition where the check occurs before the MixMonitor
thread has ended and finished writing out the file,
causing the file to be incomplete and the test to fail.

The test now waits for the 3 expected AMI events indicating
the MixMonitor thread has ended before checking the size of
the recorded files.

ASTERISK-25958 #close

Change-Id: I0b87c02b5728615755cd15a551f1cb8e40e4ad17