voicemail: Add test for passwordlocation=spooldir
Adds a test that changes a voicemail password when the option
passwordlocation=spooldir is set. The password verifies that
audio prompts are given for the options as expected and then
on shutdown the presence of secret.conf in the user's voicemail
box folder is checked and verified to contain the expected
password.

The patch also adds a small improvement to
check_voicemail_options_change_password so that it can be ran
stand-alone without the user having to create
/tmp/asterisk-testsuite prior to running the test.

ASTERISK-25917
Reported-by: Jonathan Rose

Change-Id: I996c279e98efffe18bc6df55a395490256d560b1