test_case: Add option to not end test when iterations are exhausted.
The test case module can be ended through a few mechanisms. One of
these is when all iterations are exhausted. This is problematic for
the manager redirect tests as they perform actions after the
test would normally be ended.

This change adds an option which causes the test to not be
ended when all iterations have been exhausted. It is up to
alternative ways to trigger the termination of the test.

The manager redirect dual tests have been updated to use this
option and they are ended when all expected events occur instead.

ASTERISK-25967 #close

Change-Id: Ia0a73ae5868432b102cdcde5aa54364299415a14