publish: Add tests for publishing of extension state.
This adds three tests: one for application/pidf+xml, one
for application/xpidf+xml and one for application/dialog-info+xml.
These tests are written to publish extension state to another
server and to enact device state changes which would result
in extension state changes. A SIPp scenario for each type
is written which examines the body and confirms it is
what it should be for each extension state change.

ASTERISK-26027 #close

Change-Id: I82325628ba96914f148a3697ae9bf3875ea8ef55