tests/rest_api/channels/playback: Add a basic test for media lists
This patch adds a basic test for playing back a list of media. The test
spawns a channel which has three sound files played back on it. The test
verifies the events for each sound file, as well as the expected events
for the entire play list. The test also verifies that the pause/unpause
operation works on a media file in the list.

ASTERISK-26022

Change-Id: I9a569d78d7fdb1072a52103f35230caedce40857