tests/channels/pjsip/publish: Fix presence/dialog-info+xml tests.
These tests when run on Jenkins exposed some race conditions
and behavior differences in how they run. This would cause the
tests to fail. These were also observed on other systems.

The first issue is that on some systems two PUBLISH requests may
be sent for the not in use state. This is because the custom device
state is set to not in use by default and we also set it to not
in use. Depending on the system this may or may not result in two
requests being sent. We now don't set it to not in use state
until all other states have been cycled through. This ensures that
only one initial not in use PUBLISH request goes out.

It was also observed that in some cases a state in between may
be skipped if changes were processed quickly. A delay has been
introduced to allow time for the PUBLISH request to get sent.

Change-Id: Ie12bc27e1f0f3df6204c4e92acb2e93096c067a4