tests/channels/pjsip/publish: Fix double not in use PUBLISH requests.
This change disables the pbx_ael module to stop context merging.

Context merging is a problem because the result may be two
PUBLISH requests going out with a not in use state for the extension
state. Since the SIPp scenario files require a consistent number of
requests in a given order this throws them off and causes test failure.

No more should this happen!

In practice outside of tests the above behavior is actually perfectly
acceptable. The end result is that the published request still
reflects the current extension state.

Change-Id: Iafdf05f1a57e3fef5bcdb2ed004f1b76cd09499f