bridge: Relax BackgroundDetect minimum audio.
When detecting audio using BackgroundDetect it is possible
for a re-invite to be received indicating media should be
sent to a different source and that it is coming from a
different source. In this case we will eat 3 RTP packets,
or 60ms of audio, for the probation period causing the
amount of audio received to be 18ms. As the configuration
requires 20ms it will be as if the other side never
spoke and the test will fail.

This change merely relaxes the minimum to be 15ms instead.

ASTERISK-26062 #close

Change-Id: I6b58887b6796a5e05f533c59184eb32323940ce4