tests/channels/pjsip/subscriptions/mwi: Fix incorrect value checking.
If an initial NOTIFY is received for either the mwi_aggregate
or unsolicited/mailbox_count_changes tests they will incorrectly
move on to the next MWI change but the result position will remain
as the first result. This causes each subsequent NOTIFY to be
treated as incorrect.

This change makes it so that the next MWI change only happens
if the NOTIFY in question is not the initial one.

Change-Id: Iac05ae2944ab43cc0e74025aee2c1018c31e0ba0