tests/funcs/func_talkdetect: Fix bouncing test.
The test is running afoul of a race between starting a websocket
connection and starting the test call.  When the websocket is
connected, the HTTP response is sent before the stasis application
is truly ready to accept calls.  This is not an issue in the real
world because the two events are not likely to be synchronized.

* Added a one second wait in the dialplan for the initial call
before entering the stasis testsuite application.

ASTERISK-24988

Change-Id: Ic31e8d499b1feef3c4e2c313f7f3d47621e64d7f