Testsuite: Fix SIP device state test ... again.
The previous fix to this test changed the timing of originations to be
when DeviceStateChange events were received. This works great as long as
the version of Asterisk being tested has that event. Asterisk versions
prior to 13 do not have this event.

The fix is to key off an event that is common to all versions of
Asterisk: ExtensionStatus. The logger message isn't as useful as it was
with DeviceStateChange, but the test now passes with all current
versions of Asterisk.

Change-Id: I040efcb4a2e64aa7e31143dbd626bb964f1a4025