Testsuite: Fix PJSIP show_subscriptions test.
The show_subscriptions test was failing occasionally because at
shutdown, a SUBSCRIPTION_STATE_SET event was being sent with state
TERMINATED for one of the subscriptions. This would fail the test on two
counts:

1) Since a SUBSCRIPTION_STATE_SET event beyond the maximum count was
being received, the test failed because the number of events received
exceeded the expected number.

2) Since the SUBSCRIPTION_STATE_SET event in test-config.yaml had a
requirement that the state be ACTIVE, receiving an event with state
TERMINATED resulted in a failure due to the required state being
incorrect.

The fix for this is to switch the state in the SUBSCRIPTION_STATE_SET
from being a requirement to a condition. This way, only the ACTIVE state
notifications contribute to the count. The TERMINATED states are
completely ignored and have no bearing on the test results.

Change-Id: Ib17daae0763e98fafcfabc7522d2c11a3a365220