Make tests SILK-ready.
Addition of SILK understanding to the SDP code has made it so that
Asterisk can now respond to SILK offers. This means the old single audio
stream basic test fails since it does not expect for Asterisk to respond
with SILK payloads.

This commit fixes the issue by placing a maximum version on the basic
test and creating a new version of the test with a minversion of
13.11.0. This new version is exactly the same, except that the SIPp
scenario that tries all codecs now expects for Asterisk to respond with
SILK payloads in the response.

Change-Id: I6eb3a28e0928456e4625bf17219b685b9980bdb4