Make tests SILK-ready.
Addition of SILK understanding to the SDP code has made it so that
Asterisk can now respond to SILK offers. This means the old single audio
stream basic test fails since it does not expect for Asterisk to respond
with SILK payloads.

This commit fixes the issue by splitting the existing test based on
version numbers. Once past 13.11.0, an alternate version of the
uac-all-codecs.xml file is used. In this alternate version, we expect
Asterisk to include SILK payloads in the SDP response.

Change-Id: I72fadf35d4385db0195f2ac7a65611c186e1bda3